+++ packages/atomic-elements/src/components/Banners/DismissableBanner/DismissableBanner.component.tsx
@@ -6,7 +6,6 @@ import {
   RenderBaseProps,
 } from "../../../types";
 import { useRenderProps } from "@hooks/useRenderProps";
-import { MaterialIcon } from "../../Icons/MaterialIcon";
 import { Banner, BannerVariants } from "../Banner";
 import { useTranslations } from "@hooks/useTranslations";
 
@@ -52,7 +51,7 @@ export const DismissableBanner = forwardRef<
 
   return (
     <Banner ref={ref} variant={variant} {...rest}>
-      {icon && <MaterialIcon icon={icon} variant={iconVariant} />}
+      {icon && <Banner.Icon icon={icon} variant={iconVariant} />}
       <Banner.Content>{renderProps.children}</Banner.Content>
       {onDismiss && (
         <Banner.IconButton